Change pin type

+++ hw-1/hw-1.ino
@@ -9,8 +9,8 @@ public:
     void update() const;
 
 public:
-    int inputPin; /* Analog pin */
-    int outputPin; /* PWM pin */
+    uint8_t inputPin; /* Analog pin */
+    uint8_t outputPin; /* PWM pin */
 };
 
 void LedController::initPins() const
@@ -21,8 +21,8 @@ void LedController::initPins() const
 
 void LedController::update() const
 {
-    static constexpr Tiny::Pair<int, int> INPUT_RANGE = { 0, 1023 };
-    static constexpr Tiny::Pair<int, int> OUTPUT_RANGE = { 0, 255 };
+    static constexpr Tiny::Pair<unsigned, unsigned> INPUT_RANGE = { 0, 1023 };
+    static constexpr Tiny::Pair<unsigned, unsigned> OUTPUT_RANGE = { 0, 255 };
 
     auto inputValue = analogRead(inputPin);
     auto outputValue = map(inputValue, INPUT_RANGE.first, INPUT_RANGE.second,
